+++ src/main/java/com/leverx/javacourse/seller/rating/app/service/UserService.java
@@ -1,66 +1,70 @@
 package com.leverx.javacourse.seller.rating.app.service;
 
-import com.leverx.javacourse.seller.rating.app.entity.model.Item;
 import com.leverx.javacourse.seller.rating.app.entity.model.Seller;
 import com.leverx.javacourse.seller.rating.app.entity.model.User;
+import com.leverx.javacourse.seller.rating.app.entity.model.UserRoles;
 import com.leverx.javacourse.seller.rating.app.entity.model.Visitor;
 import com.leverx.javacourse.seller.rating.app.exception.EntityNotFoundException;
+import com.leverx.javacourse.seller.rating.app.exception.UnauthorisedDataModification;
 import com.leverx.javacourse.seller.rating.app.repository.UserRepository;
-import jakarta.transaction.Transactional;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.math.BigDecimal;
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
 @Service
 public class UserService{
 
-    private UserRepository<Seller> sellerRepository;
-    private UserRepository<Visitor> visitorRepository;
-    private UserRepository<User> userRepository;
-    private ItemService itemService;
+    private final UserRepository<Seller> sellerRepository;
+    private final UserRepository<Visitor> visitorRepository;
+    private final UserRepository<User> userRepository;
 
     public UserService(UserRepository<Seller> sellerRepository, UserRepository<Visitor> visitorRepository, UserRepository<User> userRepository, ItemService itemService) {
         this.sellerRepository = sellerRepository;
         this.visitorRepository = visitorRepository;
         this.userRepository = userRepository;
-        this.itemService = itemService;
     }
 
-    @Transactional
+    @Transactional(readOnly = true)
     public User findById(Long id) {
         Optional<User> requestedUser = userRepository.findById(id);
         return requestedUser.orElseThrow(EntityNotFoundException::new);
     }
 
     @Transactional
-    public Seller saveSeller(Seller seller) {
-        return sellerRepository.save(seller);
-    }
-
-    @Transactional
-    public Visitor saveVisitor(Visitor visitor) {
-        return visitorRepository.save(visitor);
+    public User createUser(User newUser) {
+        User createdUser = null;
+        if (UserRoles.SELLER.toString().equals(newUser.getRole().toString())) {
+            createdUser = sellerRepository.save((Seller)newUser);
+        } else if (UserRoles.VISITOR.toString().equals(newUser.getRole().toString())) {
+            createdUser = visitorRepository.save((Visitor) newUser);
+        }
+        return createdUser;
     }
 
     @Transactional
     public void deleteById(Long id) {
-        userRepository.deleteById(id);
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (userDetails.getUsername().equals(findById(id).getLogin()) || userDetails.getAuthorities()
+                .stream().map(GrantedAuthority::getAuthority)
+                .anyMatch(a -> UserRoles.ADMINISTRATOR.getAuthority().equals(a))){
+            userRepository.deleteById(id);
+        }else throw new UnauthorisedDataModification();
     }
 
-    @Transactional
+    @Transactional(readOnly = true)
     public List<Seller> getAllSellers (String gameTitle, BigDecimal begin, BigDecimal end){
         return sellerRepository.findAllSellersByCriteria(gameTitle, begin, end);
     }
 
-    @Transactional
+    @Transactional(readOnly = true)
     public Optional<User> findByLogin(String login) {
         return userRepository.findByLogin(login);
     }
+
 }
